requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.15.0
evaluate>=0.4.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
mlflow>=2.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
notebook>=7.0.0
jupyter>=1.0.0